+++ src/FeatureToggles/Configuration/IpAddressesElementCollection.cs
@@ -1,19 +1,19 @@
-﻿namespace FeatureToggles.Configuration.AppConfig
+﻿namespace FeatureToggles.Configuration
 {
     using System.Collections.Generic;
     using System.Configuration;
 
-    [ConfigurationCollection(typeof(ToggleElement), AddItemName = "role", CollectionType = ConfigurationElementCollectionType.BasicMap)]
-    public class RolesElementCollection : ConfigurationElementCollection, IEnumerable<RoleElement>
+    [ConfigurationCollection(typeof(IpAddressesElement), AddItemName = "ipaddress", CollectionType = ConfigurationElementCollectionType.BasicMap)]
+    public class IpAddressesElementCollection : ConfigurationElementCollection, IEnumerable<IpAddressesElement>
     {
         /// <summary>
         /// Gets or sets a url element from the collection by index
         /// </summary>
         /// <param name="index">The index</param>
         /// <returns>The url element</returns>
-        public RoleElement this[int index]
+        public IpAddressesElement this[int index]
         {
-            get => BaseGet(index) as RoleElement;
+            get => BaseGet(index) as IpAddressesElement;
             set
             {
                 if (Count <= 0)
@@ -35,12 +35,12 @@
         /// Iterator for returning url elements from the collection
         /// </summary>
         /// <returns>A url element</returns>
-        public new IEnumerator<RoleElement> GetEnumerator()
+        public new IEnumerator<IpAddressesElement> GetEnumerator()
         {
             int count = Count;
             for (int i = 0; i < count; i++)
             {
-                yield return BaseGet(i) as RoleElement;
+                yield return BaseGet(i) as IpAddressesElement;
             }
         }
 
@@ -51,7 +51,7 @@
         /// <returns>The configuration element</returns>
         protected override ConfigurationElement CreateNewElement()
         {
-            return new RoleElement();
+            return new IpAddressesElement();
         }
 
         /// <summary>
@@ -61,7 +61,7 @@
         /// <returns>The actual instance required</returns>
         protected override object GetElementKey(ConfigurationElement element)
         {
-            return ((RoleElement)element).Name;
+            return ((IpAddressesElement)element).Value;
         }
     }
 }
