+++ src/FeatureToggles/Configuration/RoleElement.cs
@@ -1,4 +1,4 @@
-﻿namespace FeatureToggles.Configuration.AppConfig
+﻿namespace FeatureToggles.Configuration
 {
     using System.Configuration;
