+++ src/FeatureToggles/Configuration/UsersElement.cs
@@ -1,4 +1,4 @@
-﻿namespace FeatureToggles.Configuration.AppConfig
+﻿namespace FeatureToggles.Configuration
 {
     using System.Configuration;
