+++ src/FeatureToggles/Providers/AppConfigDataProvider.cs
@@ -1,30 +1,30 @@
 ﻿//-----------------------------------------------------------------------
 // <copyright file="AppConfigDataProvider.cs" company="Code Miners Limited">
 //  Copyright (c) 2019 Code Miners Limited
-//
+//   
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU Lesser General Public License as published by
 //  the Free Software Foundation, either version 3 of the License, or
 //  (at your option) any later version.
-//
+//  
 //  This program is distributed in the hope that it will be useful,
 //  but WITHOUT ANY WARRANTY; without even the implied warranty of
 //  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
 //  GNU Lesser General Public License for more details.
-//
+//  
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with this program.If not, see<https://www.gnu.org/licenses/>.
 // </copyright>
 //-----------------------------------------------------------------------
 
-namespace FeatureToggles.Providers.AppConfig
+namespace FeatureToggles.Providers
 {
     using System;
     using System.Collections.Generic;
     using System.Configuration;
     using System.Linq;
     using System.Net;
-    using FeatureToggles.Configuration.AppConfig;
+    using Configuration;
     using Models;
     using Util;
 
@@ -64,7 +64,7 @@ namespace FeatureToggles.Providers.AppConfig
         public Toggle GetFlag(string name)
         {
             if (string.IsNullOrWhiteSpace(name))
-            {
+            { 
                 Logger.Error("Fetching flag with no name provided");
                 return Toggle.Empty;
             }
