+++ src/FeatureTogglesFrameworkTests/App.config
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <configSections>
-    <section name="ToggleConfiguration" type="FeatureToggles.Configuration.AppConfig.ToggleConfigurationSection, FeatureToggles"/>
+    <section name="ToggleConfiguration" type="FeatureToggles.Configuration.ToggleConfigurationSection, FeatureToggles"/>
   </configSections>
   <appSettings>
     <add key="Toggle:Enabled" value="true"/>
