+++ src/FeatureTogglesFrameworkTests/StrongTypeTests.cs
@@ -24,9 +24,7 @@ namespace ToggleTestsApp.Config
     using System.Diagnostics.CodeAnalysis;
     using FeatureToggles;
     using FeatureToggles.Configuration;
-    using FeatureToggles.Configuration.AppConfig.Providers;
     using FeatureToggles.Providers;
-    using FeatureToggles.Providers.AppConfig;
     using Moq;
     using NUnit.Framework;
